pyproject.toml
[tool.poetry]
name = "nabeera-bareera-backend"
version = "0.1.0"
description = "FastAPI ecommerce backend"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.0"
uvicorn = {extras=["standard"], version="^0.30.0"}
motor = "^3.5.1"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
python-jose = "^3.3.0"
passlib = {extras=["bcrypt"], version="^1.7.4"}
itsdangerous = "^2.2.0"
slowapi = "^0.1.9"
cloudinary = "^1.41.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
httpx = "^0.27.0"
ruff = "^0.5.6"